requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.1.6
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
autopep8==2.0.1
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2021.5.30
cffi==1.17.1
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
commitizen==3.13.0
debugpy==1.8.6
decli==0.6.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.20.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.5.33
idna==2.10
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==8.28.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mysql-connector-python==8.0.32
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.8.0
notebook_shim==0.2.4
numpy==1.26.2
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pillow==10.4.0
platformdirs==4.1.0
pre-commit==3.6.0
prometheus_client==0.21.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.6
pure_eval==0.2.3
pyasn1==0.6.1
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.9.0
PySimpleGUI==4.60.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pywin32==307
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.2.0
questionary==2.0.1
referencing==0.35.1
requests==2.25.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
Send2Trash==1.8.3
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.3.0
tomlkit==0.12.3
tornado==6.4.1
traitlets==5.14.3
ttkbootstrap==1.10.1
typer==0.12.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
uri-template==1.3.0
urllib3==1.26.6
uvicorn==0.30.1
virtualenv==20.25.0
watchfiles==0.22.0
wcwidth==0.2.12
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
zipp==3.17.0
